backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
ultralytics==8.3.0
opencv-python-headless==4.10.0.84
filterpy==1.4.5
scipy==1.14.0
numpy==1.26.4
firebase-admin==6.6.0
google-cloud-storage==2.18.0
